requirements.txt
langchain==0.2.16
langchain-community==0.2.17
langchain-text-splitters==0.2.2


flask>=2.3
pypdf>=4.0
python-dotenv>=1.0

pinecone[grpc]>=3.0
langchain-pinecone==0.1.3

langchain-groq>=0.1.5
langchain-openai>=0.1.8
langchain-experimental>=0.0.60

sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.31.0
torch>=2.0

ipykernel>=6.29
-e .
